move dimension adjustment tests
