+++ src/app/views/common/dimensions/dimensions-adjustment.ts
