+++ src/app/views/query-runner/request/auth/Auth.spec.tsx
@@ -42,7 +42,7 @@ console.warn = jest.fn()
 // eslint-disable-next-line no-console
 console.error = jest.fn();
 
-jest.mock('../../../common/dimensions-adjustment.ts', () => {
+jest.mock('../../../common/dimensions/dimensions-adjustment.ts', () => {
   return {
     convertVhToPx: jest.fn(() => {
       return 60
