+++ src/app/views/query-runner/request/auth/Auth.tsx
@@ -11,7 +11,7 @@ import { translateMessage } from '../../../../utils/translate-messages';
 import { classNames } from '../../../classnames';
 import { trackedGenericCopy } from '../../../common/copy';
 import { CopyButton } from '../../../common/copy/CopyButton';
-import { convertVhToPx } from '../../../common/dimensions-adjustment';
+import { convertVhToPx } from '../../../common/dimensions/dimensions-adjustment';
 import { authStyles } from './Auth.styles';
 
 export function Auth(props: any) {
