+++ src/app/views/query-runner/request/headers/RequestHeaders.spec.tsx
@@ -45,7 +45,7 @@ jest.mock('react-redux', () => {
   }
 })
 
-jest.mock('../../../../app/views/common/dimensions-adjustment.ts', () => {
+jest.mock('../../../../app/views/common/dimensions/dimensions-adjustment.ts', () => {
   return {
     convertVhToPx: jest.fn(() => {
       return 60
